backend/resume_api/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
openai
